Move MultiVitalChartsView to charts folder
